fix: reason for removal date bug
