--- application/CohortManager/src/Functions/Shared/Data/Database/ICreateParticipantData.cs
@@ -1,8 +1,0 @@
-namespace Data.Database;
-
-using Model;
-
-public interface ICreateParticipantData
-{
-    public Task<bool> CreateParticipantEntry(ParticipantCsvRecord participantCsvRecord);
-}
