+++ application/CohortManager/src/Functions/Shared/Data/Database/IDatabaseHelper.cs
@@ -2,7 +2,6 @@ namespace Data.Database;
 
 public interface IDatabaseHelper
 {
-    object ParseDates(string dateString);
     object ConvertNullToDbNull(string value);
     bool CheckIfNumberNull(string property);
     int ParseExceptionFlag(object exception);
