+++ application/CohortManager/src/Functions/screeningDataServices/createParticipant/Program.cs
@@ -12,7 +12,6 @@ var host = new HostBuilder()
         .Build()
     .ConfigureServices(services =>
     {
-        services.AddTransient<ICreateParticipantData, CreateParticipantData>();
         services.AddSingleton<ICreateResponse, CreateResponse>();
         services.AddTransient<IDatabaseHelper, DatabaseHelper>();
         services.AddSingleton<ICallFunction, CallFunction>();
